+++ navigation/RELEASE_NOTES.md
@@ -3,3 +3,5 @@
 ## 0.5.0-alpha
 ##### Navigation
 * ANDDEP-312 Added navigation module
+* ANDDEP-1113 Fixed fragmentManager crash, on adding fragment when app is in background.
+* Fragment manager crash fix updated for Xiaomi on Android 9